requirements.txt
seaborn
pandas
matplotlib
requests
pyyaml